BUGFIX Fix VM handling in DRS in a case of combined license used
This patch fixes the problem with wrong handling of a license's limit on a
running VMs number in a case of a combined license.

The check in DRS is follows:
 1) compare a sum of running VMs & CTs with servers_total
 2) if the object is CT - than compare a number of running CTs with ct_limit
 3) if the object is VM - than compare a number of running VMs with nr_vms

In a case of combined license rmond reports nr_vms equal to 0, which triggers a
fail
on the 3rd check. In order to fix it, it is enough to report nr_vms ==
servers_total here. In this case a logic will be absolutely correct, because
actual check will be performed at the 1st step and will be always successful at
the 3rd one.

#PSBM-24116:fixed

Signed-off-by: Dmitry Mishin <dim@parallels.com>
Acked-by: Sergey Galas <shrike@parallels.com>
Acked-by: Pavel Emelianov <xemul@parallels.com>

Issues:
Fixes Critical Issue PSBM-24116 - "pdrs incorrectly handles combined licenses" #PSBM-24116