BUGFIX @103011 Review @103011 "Add license CT usage to node info" Reviewed-by: shrike Observers: dim * Problem:
In pdrs we use number of running CTs as number of license CT usage but it is not correct
as 1 and 50 ids are not account for usage and there could be temporary CT of number up to 2 with lifetime up to 5 hours that are not accounted too.

* Solution

Include lisence CT usage into rmond data.

#PSBM-28708
Issues:
Affects Major Issue PSBM-28708 - "PDRS: Service (1) and MN (50) containers should not be counted in running containers resource" #PSBM-28708